PCA regtest setup
