+++ DirectProgramming/C++SYCL_FPGA/ReferenceDesigns/pca/CMakeLists.txt
@@ -11,7 +11,7 @@ endif()
 
 cmake_minimum_required (VERSION 3.4)
 
-project(EIGEN CXX)
+project(PCA CXX)
 
 set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR})
 set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR})
