+++ DirectProgramming/C++SYCL_FPGA/ReferenceDesigns/pca/sample.json
@@ -1,6 +1,6 @@
 {
-  "guid": "4228581F-9DF8-4696-9B1C-0B31286B97C3",
-  "name": "EIGEN",
+  "guid": "1992581F-9DF8-4696-9B1C-0B31286B97C3",
+  "name": "PCA",
   "categories": ["Toolkit/oneAPI Direct Programming/DPC++ FPGA/Reference Designs"],
   "description": "Reference design demonstrating high-performance QR Decomposition (QRD) of real and complex matrices on a Intel® FPGA",
   "toolchain": ["dpcpp"],
